Improvements in memory efficiency (solves #438). (#460)
* add RVineStructure::needed_hfunc1/2()

* make needed_hfunc1/2 short unsigned

* simplify RVineStructure constructor while I'm at it

* remove redundant copies of vine_struct members

* non-copying trim()

* non-copying rotate_data()

* memory efficient Vinecop::pdf()

* avoid a copy in pdf_c_d

* avoid needless copies in VinecopSelector

* remove <algorithm> in tools_eigen.ipp

+++ test/src_test/include/test_rvine_structure.hpp
@@ -132,12 +132,12 @@ TEST(rvine_structure, needed_hfunc1_is_correct)
          7, 3, 0, 0, 0, 0, 0,
          4, 0, 0, 0, 0, 0, 0;
 
-  TriangularArray<size_t> true_hfunc1({ { 0, 0, 0, 0, 1, 1 },
-                                        { 0, 0, 0, 1, 1 },
-                                        { 0, 0, 0, 1 },
-                                        { 0, 0, 0 },
-                                        { 0, 1 },
-                                        { 0 } });
+  TriangularArray<short unsigned> true_hfunc1({ { 0, 0, 0, 0, 1, 1 },
+                                                { 0, 0, 0, 1, 1 },
+                                                { 0, 0, 0, 1 },
+                                                { 0, 0, 0 },
+                                                { 0, 1 },
+                                                { 0 } });
 
   RVineStructure rvine_structure(mat);
   EXPECT_EQ(rvine_structure.get_needed_hfunc1(), true_hfunc1);
@@ -157,12 +157,12 @@ TEST(rvine_structure, needed_hfunc2_is_correct)
          7, 3, 0, 0, 0, 0, 0,
          4, 0, 0, 0, 0, 0, 0;
 
-  TriangularArray<size_t> true_hfunc2({ { 1, 1, 1, 1, 1, 1 },
-                                        { 1, 1, 1, 1, 1 },
-                                        { 1, 1, 1, 1 },
-                                        { 1, 1, 1 },
-                                        { 1, 0 },
-                                        { 0 } });
+  TriangularArray<short unsigned> true_hfunc2({ { 1, 1, 1, 1, 1, 1 },
+                                                { 1, 1, 1, 1, 1 },
+                                                { 1, 1, 1, 1 },
+                                                { 1, 1, 1 },
+                                                { 1, 0 },
+                                                { 0 } });
 
   RVineStructure rvine_structure(mat);
   EXPECT_EQ(rvine_structure.get_needed_hfunc2(), true_hfunc2);
